requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
curl_cffi==0.7.4
pydantic==2.11.1
beautifulsoup4==4.13.4
python-dotenv==1.1.0
# ----------------------------------------------------------------------------
# Patchright — drop-in undetected Playwright replacement (same API surface).
# Patches CDP Runtime.enable / Console.enable / navigator.webdriver / command
# flag leaks so CreepJS & Cloudflare rarely flag us as a bot.
#
# Post-install on every deploy (VPS + Docker):
#   patchright install chromium
# (Optional, higher stealth if real Chrome is available on the box:
#   patchright install chrome   → then pass channel="chrome" at launch)
# ----------------------------------------------------------------------------
patchright==1.56.0
